Dataduct 0.2.0

+++ dataduct/s3/utils.py
@@ -31,7 +31,8 @@ def read_from_s3(s3_path):
     Returns:
         results(str): Contents of the file as a string
     """
-    assert isinstance(s3_path, S3Path), 'input path should be of type S3Path'
+    if not isinstance(s3_path, S3Path):
+        raise ETLInputError('Input path should be of type S3Path')
 
     bucket = get_s3_bucket(s3_path.bucket)
     key = bucket.get_key(s3_path.key)
@@ -47,8 +48,11 @@ def upload_to_s3(s3_path, file_name=None, file_text=None):
         file_name(str): Name of the file to be uploaded to s3
         file_text(str): Contents of the file to be uploaded
     """
-    assert isinstance(s3_path, S3Path), 'input path should be of type S3Path'
-    assert any([file_name, file_text]), 'file_name or text should be given'
+    if not isinstance(s3_path, S3Path):
+        raise ETLInputError('Input path should be of type S3Path')
+
+    if not any([file_name, file_text]):
+        raise ETLInputError('File_name or text should be given')
 
     bucket = get_s3_bucket(s3_path.bucket)
     if s3_path.is_directory:
@@ -78,8 +82,7 @@ def copy_within_s3(s3_old_path, s3_new_path, raise_when_no_exist=True):
     key = bucket.get_key(s3_old_path.key)
     if key:
         key.copy(s3_new_path.bucket, s3_new_path.key)
-
-    if raise_when_no_exist and not key:
+    elif raise_when_no_exist:
         raise ETLInputError('The key does not exist: %s' % s3_old_path.uri)
 
 
@@ -91,9 +94,14 @@ def upload_dir_to_s3(s3_path, local_path, filter_function=None):
         local_path(file_path): Input path of the file to be uploaded
         filter_function(function): Function to filter out directories
     """
-    assert isinstance(s3_path, S3Path), 'input path should be of type S3Path'
-    assert s3_path.is_directory, 'S3 path must be directory'
-    assert os.path.isdir(local_path), 'Local path must be a directory'
+    if not isinstance(s3_path, S3Path):
+        raise ETLInputError('Input path should be of type S3Path')
+
+    if not s3_path.is_directory:
+        raise ETLInputError('S3 path must be directory')
+
+    if not os.path.isdir(local_path):
+        raise ETLInputError('Local path must be a directory')
 
     bucket = get_s3_bucket(s3_path.bucket)
 
@@ -119,8 +127,11 @@ def download_dir_from_s3(s3_path, local_path):
         s3_path(S3Path): Input path of the file to be downloaded
         local_path(file_path): Output path of the file to be downloaded
     """
-    assert isinstance(s3_path, S3Path), 'input path should be of type S3Path'
-    assert s3_path.is_directory, 'S3 path must be directory'
+    if not isinstance(s3_path, S3Path):
+        raise ETLInputError('Input path should be of type S3Path')
+
+    if not s3_path.is_directory:
+        raise ETLInputError('S3 path must be directory')
 
     bucket = get_s3_bucket(s3_path.bucket)
     keys = bucket.get_all_keys(prefix=s3_path.key + '/')
@@ -146,15 +157,56 @@ def delete_dir_from_s3(s3_path):
     Args:
         s3_path(S3Path): Path of the directory to be deleted
     """
-    assert isinstance(s3_path, S3Path), 'input path should be of type S3Path'
-    assert s3_path.is_directory, 'S3 path must be directory'
+    if not isinstance(s3_path, S3Path):
+        raise ETLInputError('Input path should be of type S3Path')
+
+    if not s3_path.is_directory:
+        raise ETLInputError('S3 path must be directory')
 
     bucket = get_s3_bucket(s3_path.bucket)
     prefix = s3_path.key
 
     # Enforce this to be a folder's prefix
-    if not prefix.endswith('/'):
-        prefix += '/'
+    prefix += '/' if not prefix.endswith('/') else ''
+
     keys = bucket.get_all_keys(prefix=s3_path.key)
     for key in keys:
         key.delete()
+
+
+def copy_dir_with_s3(s3_old_path, s3_new_path, raise_when_no_exist=True):
+    """Copies files from one S3 Path to another
+
+    Args:
+        s3_old_path(S3Path): Output path of the file to be uploaded
+        s3_new_path(S3Path): Output path of the file to be uploaded
+        raise_when_no_exist(bool, optional): Raise error if file not found
+
+    Raises:
+        ETLInputError: If s3_old_path does not exist
+    """
+    if not isinstance(s3_old_path, S3Path):
+        raise ETLInputError('S3 old path should be of type S3Path')
+
+    if not s3_old_path.is_directory:
+        raise ETLInputError('S3 old path must be directory')
+
+    if not isinstance(s3_new_path, S3Path):
+        raise ETLInputError('S3 new path should be of type S3Path')
+
+    if not s3_new_path.is_directory:
+        raise ETLInputError('S3 new path must be directory')
+
+    bucket = get_s3_bucket(s3_old_path.bucket)
+    prefix = s3_old_path.key
+
+    # Enforce this to be a folder's prefix
+    prefix += '/' if not prefix.endswith('/') else ''
+
+    keys = bucket.get_all_keys(prefix=s3_old_path.key)
+    for key in keys:
+        if key:
+            key.copy(s3_new_path.bucket,
+                     os.path.join(s3_new_path.key, os.path.basename(key.key)))
+        elif raise_when_no_exist:
+            raise ETLInputError('The key does not exist: %s' % s3_old_path.uri)
